requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
suds-py3==1.4.5.0
urllib3==1.26.8
pandas~=1.4.1
python-gitlab~=3.2.0

